+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+### Fixed
+- Improper harness for KLEE, when a constrained field is an array of non-primitive cells [#97](https://github.com/OCamlPro/seacoral/pull/97)
 
 ## 1.1.0 ( 2026-03-30 )